pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <ilya.kazyamov@gmail.com>"]
repository = "https://github.com/kazyamov/python-project-50"
packages = [ { include = 'gendiff' }, ]
classifiers = [
	"Environment :: Console",
	"Programming Language :: Python"
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
flake8 = "^5.0.4"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'